Factored out progress modal into its own component
